requirements.txt
Flask==1.0.2
google-cloud-vision==0.30.1
google-cloud-storage== 1.7.0
gunicorn==19.8.1
requests==2.19.1
WTForms-Appengine==0.1
line-bot-sdk==1.8.0
